# Fix random module performance regression