--- a/cupy/random/_bit_generator.pyx
+++ b/cupy/random/_bit_generator.pyx
@@ -71,10 +71,12 @@ class BitGenerator:
 
 class _cuRANDGenerator(BitGenerator):
     # Size is the number of threads that will be initialized
-    def __init__(self, seed=None, *, size=1000*256):
+    def __init__(self, seed=None, *, size=-1):
         super().__init__(seed)
         # Raw kernel has problems with integers with the 64th bit set
         self._seed = self._seed_seq.generate_state(1, numpy.uint32)[0]
+        if size < 0:
+           size = 8 * 256 * cupy.cuda.runtime.cudaDevAttrMultiProcessorCount
         self._size = size
         cdef uint64_t b_size = self._type_size() * size
         self._state = cupy.zeros(b_size, dtype=numpy.int8)