--- a/cupy/random/_generator_api.pyx
+++ b/cupy/random/_generator_api.pyx
@@ -44,8 +44,9 @@ cdef extern from 'cupy_distributions.cuh' nogil:
         int generator, intptr_t state, ssize_t state_size, intptr_t out,
         ssize_t size, intptr_t stream, intptr_t arg1)
     void hypergeometric(
-        int generator, intptr_t state, ssize_t state_size, intptr_t out, ssize_t size,
-        intptr_t stream, intptr_t arg1, intptr_t arg2, intptr_t arg3)
+        int generator, intptr_t state, ssize_t state_size, intptr_t out,
+        ssize_t size, intptr_t stream,
+        intptr_t arg1, intptr_t arg2, intptr_t arg3)
     void logseries(
         int generator, intptr_t state, ssize_t state_size, intptr_t out,
         ssize_t size, intptr_t stream, intptr_t arg1)
@@ -67,8 +68,9 @@ cdef extern from 'cupy_distributions.cuh' nogil:
         int generator, intptr_t state, ssize_t state_size, intptr_t out,
         ssize_t size, intptr_t stream, intptr_t arg1)
     void binomial(
-        int generator, intptr_t state, ssize_t state_size, intptr_t out, ssize_t size,
-        intptr_t stream, intptr_t arg1, intptr_t arg2, intptr_t arg3)
+        int generator, intptr_t state, ssize_t state_size, intptr_t out,
+        ssize_t size, intptr_t stream,
+        intptr_t arg1, intptr_t arg2, intptr_t arg3)
 
 
 cdef _ndarray_base _array_data(_ndarray_base x):
@@ -1031,7 +1033,7 @@ def random_raw(generator, out):
     _launch_dist(generator, raw, out, ())
 
 
-cdef void _launch_dist_split(
+cdef void _launch(
         func, int generator, intptr_t state, intptr_t strm,
         int bsize, out, args):
     cdef ssize_t size = out.size
@@ -1040,44 +1042,19 @@ cdef void _launch_dist_split(
         # Avoid issues launching empty grids in CUDA 10.2
         return
     nargs = [
-        _array_data(cupy.ascontiguousarray(a))
+        _array_data(a)
         if isinstance(a, cupy.ndarray) else a for a in args]
     args_ptr = [
         <intptr_t>a.data.ptr
         if isinstance(a, cupy.ndarray) else a for a in nargs]
-    func(generator, state, bsize, <intptr_t>out.data.ptr, size, strm, *args_ptr)
-    # if size <= bsize:
-    #     nargs = [
-    #         _array_data(a)
-    #         if isinstance(a, cupy.ndarray) else a for a in args]
-    #     args_ptr = [
-    #         <intptr_t>a.data.ptr
-    #         if isinstance(a, cupy.ndarray) else a for a in nargs]
-    #     func(generator, state, bsize, <intptr_t>out.data.ptr, size, strm, *args_ptr)
-    # elif size // shape[0] <= bsize:
-    #     step = bsize // (size // shape[0])
-    #     for start in range(0, shape[0], step):
-    #         end = min(start + step, shape[0])
-    #         nargs = [
-    #             a[start:end]
-    #             if isinstance(a, cupy.ndarray) else a for a in args]
-    #         _launch_dist_split(
-    #             func, generator, state, strm, bsize, out[start:end], nargs)
-    # else:
-    #     for index in range(shape[0]):
-    #         nargs = [
-    #             a[index]
-    #             if isinstance(a, cupy.ndarray) else a for a in args]
-    #         _launch_dist_split(
-    #             func, generator, state, strm, bsize, out[index], nargs)
+
+    func(generator, state, bsize,
+         <intptr_t>out.data.ptr, size, strm, *args_ptr)
 
 
 cdef void _launch_dist(bit_generator, func, out, args) except*:
-    # The generator might only have state for a few number of threads,
-    # what we do is to split the array filling in several chunks that are
-    # generated sequentially using the same state
     cdef intptr_t strm = stream.get_current_stream_ptr()
     cdef intptr_t state = <intptr_t>bit_generator.state()
     cdef int generator = bit_generator.generator
     cdef bsize = bit_generator._state_size()
-    _launch_dist_split(func, generator, state, strm, bsize, out, args)
+    _launch(func, generator, state, strm, bsize, out, args)