@emcastillo 2023-05-10
random state was being used from global memory by curand routines making it very slow.
Caching it and updating on kernel exit resulted in much much faster ops