@leofang 2023-05-15
Would be nice to highlight the double-digit perf improvement in the release note!