@leofang 2023-05-25
Q: @asi1024 @kmaehashi would it be possible to backport this PR, since it counts as a kind of fix (for perf)?